+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.6.0
+
+- Add `ShadTabs` component
+- Add `ShadColorScheme.fromName` to easily create a color scheme from a name (String)
+- Add `package` to `ShadImage` (thanks to @farhanfadila1717)
+- Fix `decoration` of form fields
+- Fix selection controls of `ShadInput`
+
 ## 0.5.7
 
 - Renamed the breakpoints